Level 2 Scene
Issue currently: Whenever I use the tether I get a "Could not find path to (96.6, 214.9, 0.0) error in the Debug Log. Also when I tether back again a second time with a keycode already in my inventory, the keycode door will not show the "Press E" object to open the door.