Fixed Stasis Bubble Add/Remove Interactions
Fixed some problems with adding and removing stasis bubbles. If the mouse is over overlapping stasis bubbles, only the top one will be removed when you right click. Also, the code no longer prevents you from shooting a stasis bubble when the mouse is over a tether point or a stasis bubble