Tether Placement Bug Fix
Added a check in the TetherManager that prevents you from placing another tether point while the current tether place animation is playing. Possibly as a result of the animation event change, this led to a bug in which rapidly tapping space to place all tether points might result in unremovable tether points. This should no longer happen.